Final time: 1 hour 38 minutes. I Quit
I simply did not find this engine to be enjoyable to use. I have used Bevy in the past, and absolutely love it. This engine just gives me headaches. Built in functions just do not work. The documentation does not provide proper info, and I don't want to watch a 18 part tutorial series just to have a single thing work.